requirements.txt
streamlit
openai
SpeechRecognition
gTTS
python-dotenv
